+++ CHANGELOG.md
@@ -27,3 +27,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [Unreleased]: https://gitlab.com/tozd/go/errors/-/compare/v0.2.0...main
 [0.2.0]: https://gitlab.com/tozd/go/errors/-/compare/v0.1.0...v0.2.0
 [0.1.0]: https://gitlab.com/tozd/go/errors/-/tags/v0.1.0
+
+<!-- markdownlint-disable-file MD024 -->